+++ packages/extensions/modelerfour/src/modeler/modelerfour.ts
@@ -76,6 +76,7 @@ import { Session, Channel } from "@autorest/extension-base";
 import { Interpretations, XMSEnum } from "./interpretations";
 import { fail, minimum, pascalCase, knownMediaType, KnownMediaType } from "@azure-tools/codegen";
 import { ModelerFourOptions } from "./modelerfour-options";
+import { isContentTypeParameterDefined } from "./utils";
 
 /** adds only if the item is not in the collection already
  *
@@ -1367,9 +1368,11 @@ export class ModelerFour {
         http,
       },
     });
-    this.session.log(`Options ${JSON.stringify(this.options)}`, {});
-    this.session.log(`Accept-param ${this.options["always-create-accept-parameter"]}`, {});
-    if (this.options[`always-create-content-type-parameter`] === true || http.mediaTypes.length > 1) {
+
+    const shouldIncludeContentType =
+      this.options[`always-create-content-type-parameter`] === true || http.mediaTypes.length > 1;
+
+    if (!isContentTypeParameterDefined(operation) && shouldIncludeContentType) {
       const scs = this.getContentTypeParameterSchema(http);
 
       // add the parameter for the binary upload.
@@ -1446,7 +1449,7 @@ export class ModelerFour {
       },
     });
 
-    if (this.options[`always-create-content-type-parameter`] === true) {
+    if (!isContentTypeParameterDefined(operation) && this.options[`always-create-content-type-parameter`] === true) {
       const scs = this.getContentTypeParameterSchema(http, true);
 
       // add the parameter for the binary upload.
@@ -1654,13 +1657,7 @@ export class ModelerFour {
             // scenario 3 : single parameterized value
 
             for (const { key: variableName, value: variable } of items(server.variables).where((each) => !!each.key)) {
-              const sch = variable.enum
-                ? this.processChoiceSchema(variableName, <OpenAPI.Schema>{
-                    type: "string",
-                    enum: variable.enum,
-                    description: variable.description || `${variableName} - server parameter`,
-                  })
-                : this.stringSchema;
+              const sch = this.getServerVariableSchema(variableName, variable);
 
               const clientdefault = variable.default ? variable.default : undefined;
 
@@ -1765,6 +1762,23 @@ export class ModelerFour {
     return baseUri;
   }
 
+  private getServerVariableSchema(variableName: string, variable: OpenAPI.ServerVariable) {
+    if (variable.enum) {
+      return this.processChoiceSchema(variableName, <OpenAPI.Schema>{
+        type: "string",
+        enum: variable.enum,
+        description: variable.description || `${variableName} - server parameter`,
+      });
+    }
+
+    if (variable["x-format"]) {
+      return this.processSchema(`${variableName}`, {
+        type: JsonType.String,
+        format: variable["x-format"],
+      });
+    }
+    return this.stringSchema;
+  }
   processApiVersionParameterForProfile() {
     throw new Error("Profile Support for API Verison Parameters not implemented.");
   }
@@ -1903,11 +1917,15 @@ export class ModelerFour {
           this.trackSchemaUsage(parameterSchema, { usage: [SchemaContext.Input] });
 
           if (parameter.in === ParameterLocation.Header && "x-ms-header-collection-prefix" in parameter) {
-            parameterSchema = new DictionarySchema(
-              parameterSchema.language.default.name,
-              parameterSchema.language.default.description,
-              parameterSchema,
+            const dictionarySchema = this.codeModel.schemas.add(
+              new DictionarySchema(
+                parameterSchema.language.default.name,
+                parameterSchema.language.default.description,
+                parameterSchema,
+              ),
             );
+            this.trackSchemaUsage(dictionarySchema, { usage: [SchemaContext.Input] });
+            parameterSchema = dictionarySchema;
           }
 
           /* regular, everyday parameter */
@@ -2103,7 +2121,11 @@ export class ModelerFour {
       this.use(header.schema, (_name, sch) => {
         let hsch = this.processSchema(this.interpret.getName(headerName, sch), sch);
         if ("x-ms-header-collection-prefix" in header) {
-          hsch = new DictionarySchema(hsch.language.default.name, hsch.language.default.description, hsch);
+          const newSchema = new DictionarySchema(hsch.language.default.name, hsch.language.default.description, hsch);
+          newSchema.language.default.header = headerName;
+          const dictionarySchema = this.codeModel.schemas.add(newSchema);
+          this.trackSchemaUsage(dictionarySchema, { usage: [SchemaContext.Input] });
+          hsch = dictionarySchema;
         }
 
         hsch.language.default.header = headerName;