+++ packages/libs/common/src/merging/merging.ts
@@ -126,14 +126,10 @@ export function resolveRValue(
 
     // resolve macro values for array values
     if (value instanceof Array) {
-      const result = new Array<any>();
-      for (const each of value) {
-        // since we're not naming the parameter,
-        // if there isn't a higher priority,
-        // we can fall back to a wide-lookup in lowerPriority.
-        result.push(resolveRValue(each, "", higherPriority || lowerPriority, null));
-      }
-      return result;
+      // since we're not naming the parameter,
+      // if there isn't a higher priority,
+      // we can fall back to a wide-lookup in lowerPriority.
+      return value.map((x) => resolveRValue(x, "", higherPriority || lowerPriority, null));
     }
   }
 
@@ -144,15 +140,34 @@ export function resolveRValue(
   return value;
 }
 
+export type ArrayMergingStrategy = "high-pri-first" | "low-pri-first";
+
+export interface MergeOptions {
+  interpolationContext?: any;
+  arrayMergeStrategy?: ArrayMergingStrategy;
+  concatListPathFilter?: (path: JsonPath) => boolean;
+}
+
+const defaultOptions: Omit<Required<MergeOptions>, "interpolationContext"> = {
+  arrayMergeStrategy: "high-pri-first",
+  concatListPathFilter: () => false,
+};
+
 export function mergeOverwriteOrAppend(
   higherPriority: any,
   lowerPriority: any,
-  concatListPathFilter: (path: JsonPath) => boolean = (_) => false,
+  options: MergeOptions = {},
   path: JsonPath = [],
 ): any {
   if (higherPriority === null || lowerPriority === null) {
     return null;
   }
+
+  const computedOptions = {
+    ...defaultOptions,
+    ...options,
+    interpolationContext: options.interpolationContext ?? higherPriority,
+  };
 
   // scalars/arrays involved
   if (
@@ -161,11 +176,7 @@ export function mergeOverwriteOrAppend(
     typeof lowerPriority !== "object" ||
     lowerPriority instanceof Array
   ) {
-    if (!(higherPriority instanceof Array) && !(lowerPriority instanceof Array) && !concatListPathFilter(path)) {
-      return higherPriority;
-    }
-
-    return [...new Set((higherPriority instanceof Array ? higherPriority : [higherPriority]).concat(lowerPriority))];
+    return mergeArray(higherPriority, lowerPriority, path, computedOptions);
   }
 
   // object nodes - iterate all members
@@ -181,20 +192,45 @@ export function mergeOverwriteOrAppend(
 
     // forward if only present in one of the nodes
     if (higherPriority[key] === undefined) {
-      result[key] = resolveRValue(lowerPriority[key], key, higherPriority, lowerPriority);
+      result[key] = resolveRValue(lowerPriority[key], key, computedOptions.interpolationContext, lowerPriority);
       continue;
     }
     if (lowerPriority[key] === undefined) {
-      result[key] = resolveRValue(higherPriority[key], key, null, higherPriority);
+      result[key] = resolveRValue(higherPriority[key], key, null, computedOptions.interpolationContext);
       continue;
     }
 
     // try merge objects otherwise
-    const aMember = resolveRValue(higherPriority[key], key, lowerPriority, higherPriority);
-    const bMember = resolveRValue(lowerPriority[key], key, higherPriority, lowerPriority);
-    result[key] = mergeOverwriteOrAppend(aMember, bMember, concatListPathFilter, subpath);
+    const aMember = resolveRValue(higherPriority[key], key, lowerPriority, computedOptions.interpolationContext);
+    const bMember = resolveRValue(lowerPriority[key], key, computedOptions.interpolationContext, lowerPriority);
+    result[key] = mergeOverwriteOrAppend(
+      aMember,
+      bMember,
+      { ...computedOptions, interpolationContext: computedOptions.interpolationContext[key] },
+      subpath,
+    );
   }
   return result;
+}
+
+function mergeArray(
+  higherPriority: unknown,
+  lowerPriority: unknown,
+  path: JsonPath,
+  { concatListPathFilter, arrayMergeStrategy }: Required<MergeOptions>,
+) {
+  if (!(higherPriority instanceof Array) && !(lowerPriority instanceof Array) && !concatListPathFilter(path)) {
+    return higherPriority;
+  }
+
+  const higherPriorityArray = higherPriority instanceof Array ? higherPriority : [higherPriority];
+  const lowerPriorityArray = lowerPriority instanceof Array ? lowerPriority : [lowerPriority];
+
+  if (arrayMergeStrategy === "high-pri-first") {
+    return [...new Set(higherPriorityArray.concat(lowerPriority))];
+  } else {
+    return [...new Set(lowerPriorityArray.concat(higherPriority))];
+  }
 }
 
 export function identitySourceMapping(sourceYamlFileName: string, sourceYamlAst: YAMLNode): Mapping[] {