+++ packages/libs/configuration/src/configuration-manager/configuration-manager.test.ts
@@ -43,18 +43,37 @@ describe("ConfigurationManager", () => {
       expect(output["output-folder"]).toEqual("generated-1");
     });
 
-    it("combines array values", async () => {
+    it("combines array values in order of higher priority", async () => {
       expect(output["api-version"]).toEqual(["version-1", "version-2"]);
     });
 
     it("adds value from 2nd config if not present in 1st", async () => {
       expect(output["base-folder"]).toEqual("base-folder-2");
     });
+
+    it("combine nested objects", async () => {
+      await manager.addConfig({
+        "use-extension": {
+          "@autorest/csharp": "latest",
+        },
+      });
+      await manager.addConfig({
+        "use-extension": {
+          "@autorest/modelerfour": "latest",
+        },
+      });
+
+      const output = await manager.resolveConfig();
+      expect(output["use-extension"]).toEqual({
+        "@autorest/csharp": "latest",
+        "@autorest/modelerfour": "latest",
+      });
+    });
   });
 
   describe("adding a single file with multiple blocks", () => {
     it("override values with blocks defined later", async () => {
-      await manager.addConfigFile({
+      manager.addConfigFile({
         type: "file",
         fullPath: "/dev/path/readme.md",
         configs: [
@@ -77,11 +96,38 @@ describe("ConfigurationManager", () => {
       const output = await manager.resolveConfig();
       expect(output["output-folder"]).toEqual("generated-2");
       expect(output["base-folder"]).toEqual("base-folder-1");
-      expect(output["api-version"]).toEqual(["version-2", "version-1"]);
+      expect(output["api-version"]).toEqual(["version-1", "version-2"]);
+    });
+
+    it("merges array in the order they are defined", async () => {
+      manager.addConfigFile({
+        type: "file",
+        fullPath: "/dev/path/readme.md",
+        configs: [
+          {
+            config: {
+              "api-version": ["version-1"],
+            },
+          },
+          {
+            config: {
+              "api-version": ["version-2"],
+            },
+          },
+          {
+            config: {
+              "api-version": ["version-3"],
+            },
+          },
+        ],
+      });
+
+      const output = await manager.resolveConfig();
+      expect(output["api-version"]).toEqual(["version-1", "version-2", "version-3"]);
     });
 
     it("use condition from previous blocks", async () => {
-      await manager.addConfigFile({
+      manager.addConfigFile({
         type: "file",
         fullPath: "/dev/path/readme.md",
         configs: [
@@ -111,7 +157,112 @@ describe("ConfigurationManager", () => {
       const output = await manager.resolveConfig();
       expect(output["output-folder"]).toEqual("generated-1");
       expect(output["base-folder"]).toEqual("base-folder-1");
-      expect(output["api-version"]).toEqual(["version-3", "version-1"]);
+      expect(output["api-version"]).toEqual(["version-1", "version-3"]);
+    });
+  });
+
+  describe("interpolate previous values", () => {
+    it("interpolate from previous config", async () => {
+      await manager.addConfig({
+        name: "FooBar",
+      });
+      await manager.addConfig({
+        namespace: "$(name).Client",
+      });
+
+      const output = await manager.resolveConfig();
+      expect(output["namespace"]).toEqual("FooBar.Client");
+    });
+
+    it("interpolate from same config but defined before", async () => {
+      await manager.addConfig({
+        name: "FooBar",
+        namespace: "$(name).Client",
+      });
+
+      const output = await manager.resolveConfig();
+      expect(output["namespace"]).toEqual("FooBar.Client");
+    });
+
+    it("interpolate with higher priority value(defined before) instead of the one in the same block", async () => {
+      await manager.addConfig({
+        name: "FooBarOverride",
+      });
+      await manager.addConfig({
+        name: "FooBar",
+        namespace: "$(name).Client",
+      });
+
+      const output = await manager.resolveConfig();
+      expect(output["namespace"]).toEqual("FooBarOverride.Client");
+    });
+
+    it("interpolate with higher priority value(defined before) instead of the one in the same block", async () => {
+      await manager.addConfig({
+        name: "FooBarOverride",
+      });
+
+      manager.addConfigFile({
+        type: "file",
+        fullPath: "/dev/path/readme.md",
+        configs: [
+          {
+            config: {
+              name: "FooBar",
+              namespace: "$(name).Client",
+            },
+          },
+        ],
+      });
+      const output = await manager.resolveConfig();
+      expect(output["namespace"]).toEqual("FooBarOverride.Client");
+    });
+
+    it("interpolate from the last block", async () => {
+      manager.addConfigFile({
+        type: "file",
+        fullPath: "/dev/path/readme.md",
+        configs: [
+          {
+            config: {
+              name: "FooBar",
+            },
+          },
+          {
+            config: {
+              name: "FooBarOverride",
+              namespace: "$(name).Client",
+            },
+          },
+        ],
+      });
+      const output = await manager.resolveConfig();
+      expect(output["namespace"]).toEqual("FooBarOverride.Client");
+    });
+
+    it("interpolate from the last block", async () => {
+      await manager.addConfig({
+        name: "FooBarCLIOverride",
+      });
+      manager.addConfigFile({
+        type: "file",
+        fullPath: "/dev/path/readme.md",
+        configs: [
+          {
+            config: {
+              name: "FooBar",
+            },
+          },
+          {
+            config: {
+              name: "FooNextBlockOverride",
+              namespace: "$(name).Client",
+            },
+          },
+        ],
+      });
+      const output = await manager.resolveConfig();
+      expect(output["namespace"]).toEqual("FooBarCLIOverride.Client");
     });
   });
 });