+++ packages/libs/extension/src/system-requirements/python.test.ts
@@ -1,4 +1,4 @@
-import { resolvePythonRequirement } from "./python";
+import { patchPythonPath, PythonCommandLine, resolvePythonRequirement, updatePythonPath } from "./python";
 import { execute } from "../exec-cmd";
 
 jest.mock("../exec-cmd");
@@ -80,6 +80,73 @@ describe("Python system requirement", () => {
     expect(result).toEqual({
       name: "python",
       command: "python3",
+    });
+  });
+
+  describe("patchPythonPath", () => {
+    it("stub with the only compatible python version", async () => {
+      mockInstalledPython({
+        python3: "3.7.0",
+        python: "2.7.0",
+      });
+      const input: PythonCommandLine = ["python", "-c", "print();"];
+      const result = await patchPythonPath(input, { version: ">=3.6" });
+      expect(result).toEqual(["python3", "-c", "print();"]);
+    });
+
+    it("stub with the first compatible python version", async () => {
+      mockInstalledPython({
+        python3: "3.7.0",
+        python: "3.9.0",
+      });
+      const input: PythonCommandLine = ["python", "-c", "print();"];
+      const result = await patchPythonPath(input, { version: ">=3.6" });
+      expect(result).toEqual(["python3", "-c", "print();"]);
+    });
+
+    it("stub with the first compatible python version", async () => {
+      mockInstalledPython({
+        python3: "2.0.0",
+        python: "3.9.0",
+      });
+      const input: PythonCommandLine = ["python", "-c", "print();"];
+      const result = await patchPythonPath(input, { version: ">=3.6" });
+      expect(result).toEqual(["python", "-c", "print();"]);
+    });
+  });
+
+  describe("updatePythonPath (Deprecated)", () => {
+    it("stub with the only compatible python version", async () => {
+      mockInstalledPython({
+        python3: "3.7.0",
+        python: "2.7.0",
+      });
+      const input: PythonCommandLine = ["python", "-c", "print();"];
+      const result = await updatePythonPath(input);
+      expect(input).toEqual(["python3", "-c", "print();"]);
+      expect(result).toEqual(["python3", "-c", "print();"]);
+    });
+
+    it("stub with the first compatible python version", async () => {
+      mockInstalledPython({
+        python3: "3.7.0",
+        python: "3.9.0",
+      });
+      const input: PythonCommandLine = ["python", "-c", "print();"];
+      const result = await updatePythonPath(input);
+      expect(input).toEqual(["python3", "-c", "print();"]);
+      expect(result).toEqual(["python3", "-c", "print();"]);
+    });
+
+    it("stub with the first compatible python version", async () => {
+      mockInstalledPython({
+        python3: "2.0.0",
+        python: "3.9.0",
+      });
+      const input: PythonCommandLine = ["python", "-c", "print();"];
+      const result = await updatePythonPath(input);
+      expect(input).toEqual(["python", "-c", "print();"]);
+      expect(result).toEqual(["python", "-c", "print();"]);
     });
   });
 });