+++ packages/libs/oai2-to-oai3/test/scenarios/scenarios.test.ts
@@ -49,4 +49,9 @@ describe("Scenario testings", () => {
     // The expected result is the parmaeter to be included/expanded in the OpenAPI3 server property.
     await expectInputsMatchSnapshots("cross-file-parameterized-host-refs", ["swagger.json", "other.json"]);
   });
+
+  it("Convert parameterized host parameters to server variables", async () => {
+    // The expected result is the parmaeter to be included/expanded in the OpenAPI3 server property.
+    await expectInputsMatchSnapshots("parameterized-host-parameters", ["swagger.json"]);
+  });
 });